requirements.txt
pytest == 7.4.3
selene == 2.0.0rc9
allure-pytest == 2.13.2
python-dotenv == 1.0.0
requests~=2.31.0
jsonschema~=4.21.1
pydantic~=2.6.4